# petstore/src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: Petstore API
  version: '@project.version@'
servers:
  - url: 'http://localhost:8080'
    description: Petstore provider
paths:
  '/dogs/{id}':
    get:
      operationId: getDog
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: dog found
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Dog'
        '404':
          description: no dog found
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Error'
  /dogs:
    post:
      operationId: saveDog
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/Dog'
      responses:
        '201':
          description: dog added
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Dog'
  '/cats/{id}':
    get:
      operationId: getCat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: cat found
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Cat'
        '404':
          description: default response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Error'
  /cats:
    post:
      operationId: save
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/Cat'
      responses:
        '201':
          description: default response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Cat'
components:
  schemas:
    Dog:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        type:
          type: string
    Cat:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        type:
          type: string
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        path:
          type: string
        status:
          type: integer
          format: int64
        timestamp:
          type: string
          format: date-time